Move shared components.
