+++ src/article/shared/EntityEditor.js
